PHPStan Level 3 Compliance (#8118)
* Phpstan test

* Phpstan test

* Phpstan test

* [Tests] Documents CRUD operations, editables, inheritance, .... #8030

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Update lib/Targeting/Storage/SessionStorage.php


* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Update doc/Development_Documentation/23_Installation_and_Upgrade/09_Upgrade_Notes/README.md


* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Update phpstan.neon

* fixed #8261 - AbstractOrderList:getItems doesn't match interface

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* fixed u

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Update bundles/EcommerceFrameworkBundle/FilterService/FilterType/ElasticSearch/Input.php


* Update bundles/EcommerceFrameworkBundle/FilterService/FilterType/ElasticSearch/MultiSelect.php


* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Phpstan test

* Update MultiSelect.php

* Phpstan test

+++ bundles/EcommerceFrameworkBundle/VoucherService/TokenManager/Pattern.php
@@ -264,10 +264,10 @@ class Pattern extends AbstractTokenManager implements ExportableTokenManagerInte
         $prefix = $this->configuration->getPrefix();
         if (!empty($separator)) {
             if (!empty($prefix)) {
-                return strlen($this->configuration->getPrefix()) + 1 + floor($this->configuration->getLength() / $separatorCount) + $this->configuration->getLength();
+                return strlen($this->configuration->getPrefix()) + 1 + (int) floor($this->configuration->getLength() / $separatorCount) + $this->configuration->getLength();
             }
 
-            return floor($this->configuration->getLength() / $separatorCount) + $this->configuration->getLength();
+            return (int) floor($this->configuration->getLength() / $separatorCount) + $this->configuration->getLength();
         }
 
         return strlen($this->configuration->getPrefix()) + $this->configuration->getLength();
